refactor: 인터페이스 생성에 따른 클래스명 변경
