+++ src/main/java/racingcar/service/GameService.java
@@ -2,40 +2,16 @@ package racingcar.service;
 
 import java.util.List;
 
-import racingcar.domain.car.CarCollection;
-import racingcar.domain.round.Round;
-import racingcar.service.picker.NumberPicker;
+public interface GameService {
 
-public class GameService {
+	void initCarNames(List<String> carNames);
 
-	private final NumberPicker numberPicker;
-	private CarCollection carCollection;
-	private Round round;
+	void initRound(int count);
 
-	public GameService(NumberPicker numberPicker) {
-		this.numberPicker = numberPicker;
-	}
+	boolean isContinuable();
 
-	public void initCarNames(List<String> carNames) {
-		this.carCollection = new CarCollection(carNames);
-	}
+	List<String> playRound();
 
-	public void initRound(int count) {
-		this.round = new Round(count);
-	}
-
-	public boolean isContinuable() {
-		return (round.isNotFinished());
-	}
-
-	public List<String> playRound() {
-		carCollection.goForwardOrStop(numberPicker);
-		round.decreaseCount();
-		return carCollection.getStatuses();
-	}
-
-	public List<String> getWinnerNames() {
-		return carCollection.getWinnerNames();
-	}
+	List<String> getWinnerNames();
 
 }
